Move 'section.js' inside 'components' folder
